# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/learning/lessons/{lessonId}/start": {
      "post": {
        "operationId": "LessonActivityController_startLesson",
        "parameters": [
          {
            "$ref": "#/components/parameters/RequestId"
          },
          {
            "name": "lessonId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Replaying a key with the same request returns the original result.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmptyLessonActivityWriteDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {},
                    "meta": {
                      "$ref": "#/components/schemas/ApiResponseMetaDto"
                    }
                  }
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ErrorEnvelope"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "tags": [
          "LessonActivity"
        ]
      }
    },
    "/api/v1/learning/lessons/{lessonId}/activity": {
      "get": {
        "operationId": "LessonActivityController_getLessonActivity",
        "parameters": [
          {
            "$ref": "#/components/parameters/RequestId"
          },
          {
            "name": "lessonId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {},
                    "meta": {
                      "$ref": "#/components/schemas/ApiResponseMetaDto"
                    }
                  }
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ErrorEnvelope"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "tags": [
          "LessonActivity"
        ]
      }
    },
    "/api/v1/learning/lessons/{lessonId}/complete": {
      "post": {
        "operationId": "LessonActivityController_completeLesson",
        "parameters": [
          {
            "$ref": "#/components/parameters/RequestId"
          },
          {
            "name": "lessonId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Replaying a key with the same request returns the original result.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmptyLessonActivityWriteDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {},
                    "meta": {
                      "$ref": "#/components/schemas/ApiResponseMetaDto"
                    }
                  }
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ErrorEnvelope"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "tags": [
          "LessonActivity"
        ]
      }
    },
    "/api/v1/learning/topics/{topicId}/start": {
      "post": {
        "operationId": "LessonActivityController_startTopic",
        "parameters": [
          {
            "$ref": "#/components/parameters/RequestId"
          },
          {
            "name": "topicId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Replaying a key with the same request returns the original result.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmptyLessonActivityWriteDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {},
                    "meta": {
                      "$ref": "#/components/schemas/ApiResponseMetaDto"
                    }
                  }
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ErrorEnvelope"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "tags": [
          "LessonActivity"
        ]
      }
    },
    "/api/v1/learning/topics/{topicId}/complete": {
      "post": {
        "operationId": "LessonActivityController_completeTopic",
        "parameters": [
          {
            "$ref": "#/components/parameters/RequestId"
          },
          {
            "name": "topicId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Replaying a key with the same request returns the original result.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmptyLessonActivityWriteDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {},
                    "meta": {
                      "$ref": "#/components/schemas/ApiResponseMetaDto"
                    }
                  }
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ErrorEnvelope"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "tags": [
          "LessonActivity"
        ]
      }
    },
    "/api/v1/learning/exercises/{exerciseId}/attempts": {
      "post": {
        "operationId": "LessonActivityController_submitAttempt",
        "parameters": [
          {
            "$ref": "#/components/parameters/RequestId"
          },
          {
            "name": "exerciseId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Replaying a key with the same request returns the original result.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitLessonExerciseAttemptDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {},
                    "meta": {
                      "$ref": "#/components/schemas/ApiResponseMetaDto"
                    }
                  }
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ErrorEnvelope"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "tags": [
          "LessonActivity"
        ]
      },
      "get": {
        "operationId": "LessonActivityController_getAttempts",
        "parameters": [
          {
            "$ref": "#/components/parameters/RequestId"
          },
          {
            "name": "exerciseId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {},
                    "meta": {
                      "$ref": "#/components/schemas/ApiResponseMetaDto"
                    }
                  }
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ErrorEnvelope"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "tags": [
          "LessonActivity"
        ]
      }
    },
...
  },
  "info": {
    "title": "HSK System API",
    "description": "Generated from the backend code. Envelope, errors and request id: docs/api/api.md §1.",
    "version": "1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "EmptyLessonActivityWriteDto": {
        "type": "object",
        "properties": {}
      },
      "SubmitLessonExerciseAttemptDto": {
        "type": "object",
        "properties": {
          "answer": {
            "type": "object",
            "additionalProperties": true
          },
          "durationSeconds": {
            "type": "number",
            "minimum": 0,
            "maximum": 86400
          }
        },
        "required": [
          "answer"
        ]
      },
...
      "ApiResponseMetaDto": {
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "requestId": {
            "type": "string",
            "description": "Same value as the `X-Request-ID` response header."
          }
        },
        "required": [
          "timestamp",
          "requestId"
        ]
      },
...
      "ApiErrorDto": {
        "type": "object",
        "properties": {
          "details": {
            "type": "object",
            "additionalProperties": true
          },
          "code": {
            "type": "string",
            "description": "Stable machine-readable code, for example `UNAUTHORIZED`."
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
      "ApiErrorEnvelopeDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              false
            ]
          },
          "error": {
            "$ref": "#/components/schemas/ApiErrorDto"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiResponseMetaDto"
          }
        },
        "required": [
          "success",
          "error",
          "meta"
        ]
      }
    },
    "parameters": {
      "RequestId": {
        "name": "X-Request-ID",
        "in": "header",
        "required": false,
        "description": "UUID or 32 hex characters; any other value is replaced by a new UUID. Echoed in the X-Request-ID response header and meta.requestId.",
        "schema": {
          "type": "string"
        }
      }
    },
    "responses": {
      "ErrorEnvelope": {
        "description": "Error envelope.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiErrorEnvelopeDto"
            }
          }
        }
      }
    }
  }
}